requirements.txt
fastapi
uvicorn
boto3
python-multipart
jinja2
mysql-connector-python

